New namespace: "Toubab.Beinder.Extend" with all "custom" extension interfaces
Left out IPathParser, because it is not "automatically discoverable".

All extension points that should be automagically discoverable (in the near
future) are in the Toubab.Beinder.Extend namespace, and have a name starting
with "Custom". With this in place, we will be able scan for extension points
that the user (developer) *wants* to be auto-discovered, and leave alone those
extension points that the user (developer) wants to treat specially.