Change IEvent.Broadcast to a more simple listener registration
This is the first step to move towards more separation between IBindable
and the actual object it represents a property on.

This separation is needed because at this moment, the WeakReference is
to IBindable; which is not good, because in some cases IBindable will
be collected because there are no references to it any more.

In the foreseeable future, the IBindable instances will be referenced
with a hard reference, whereas the objects that are tied to will be
weakly referenced. The object will only be temporarily attached to the
IBindable, when needed.