Let StateValve replace event's payload instead of letting the IProperty do it.
This means any event emanating from an IProperty is interpreted as a
value change of that property. The payload is discarded and replaced
with the current value(s) of the property.

This subtle change means we are one step closer to "transiently" attach
and detach IBindables to objects: the object will only need to be attached
if and when an event is being handled.